backend/requirements.txt
fastapi
uvicorn
openai
sentence-transformers
pinecone-client
PyPDF2
python-dotenv
starlette
python-multipart
uvicorn[standard]
websockets
wsproto
redis